pyproject.toml
[tool.poetry]
name = "corpus-clavicle"
version = "0.1.1"
description = "Total proteomics differential analysis models and auxilaries for Corpus"
authors = ["Toan Phung <toan.phung@proteo.info>"]
readme = "README.md"
packages = [
    { include = "clavicle" },
]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.6"
django-postgres-extra = "^2.0.8"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
drf-url-filters = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
